backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
newspaper3k==0.2.8
textblob==0.17.1
scikit-learn==1.3.2
nltk==3.8.1
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
